Moving src files into a service folder. Npm test working and servlerles deploy also working
