+++ package.json
@@ -25,7 +25,7 @@
   },
   "scripts": {
     "env": "env",
-    "lint": "./node_modules/.bin/eslint **/*.js",
+    "lint": "./node_modules/.bin/eslint **/*Service/*.js",
     "pretest": "npm run lint",
     "test": "mocha test/",
     "version:patch": "npm version patch",
